+++ address.go
@@ -1,0 +1,63 @@
+package dcrlibwallet
+
+import (
+	"github.com/decred/dcrd/dcrutil"
+	"github.com/raedahgroup/dcrlibwallet/addresshelper"
+)
+
+// AddressInfo holds information about an address
+// If the address belongs to the querying wallet, IsMine will be true and the AccountNumber and AccountName values will be populated
+type AddressInfo struct {
+	Address       string
+	IsMine        bool
+	AccountNumber uint32
+	AccountName   string
+}
+
+func (lw *LibWallet) IsAddressValid(address string) bool {
+	_, err := addresshelper.DecodeForNetwork(address, lw.activeNet.Params)
+	return err == nil
+}
+
+func (lw *LibWallet) HaveAddress(address string) bool {
+	addr, err := addresshelper.DecodeForNetwork(address, lw.activeNet.Params)
+	if err != nil {
+		return false
+	}
+	have, err := lw.wallet.HaveAddress(addr)
+	if err != nil {
+		return false
+	}
+
+	return have
+}
+
+func (lw *LibWallet) AccountOfAddress(address string) string {
+	addr, err := dcrutil.DecodeAddress(address)
+	if err != nil {
+		return err.Error()
+	}
+	info, _ := lw.wallet.AddressInfo(addr)
+	return lw.AccountName(info.Account())
+}
+
+func (lw *LibWallet) AddressInfo(address string) (*AddressInfo, error) {
+	addr, err := dcrutil.DecodeAddress(address)
+	if err != nil {
+		log.Error(err)
+		return nil, err
+	}
+
+	addressInfo := &AddressInfo{
+		Address: address,
+	}
+
+	info, _ := lw.wallet.AddressInfo(addr)
+	if info != nil {
+		addressInfo.IsMine = true
+		addressInfo.AccountNumber = info.Account()
+		addressInfo.AccountName = lw.AccountName(info.Account())
+	}
+
+	return addressInfo, nil
+}
