+++ txauthor.go
@@ -1,0 +1,211 @@
+package dcrlibwallet
+
+import (
+	"bytes"
+	"context"
+	"time"
+
+	"github.com/decred/dcrd/dcrutil"
+	"github.com/decred/dcrd/txscript"
+	"github.com/decred/dcrd/wire"
+	"github.com/decred/dcrwallet/errors"
+	"github.com/decred/dcrwallet/wallet"
+	"github.com/decred/dcrwallet/wallet/txauthor"
+	"github.com/decred/dcrwallet/wallet/txrules"
+	"github.com/raedahgroup/dcrlibwallet/txhelper"
+)
+
+func (lw *LibWallet) ConstructTransaction(destAddr string, amount int64, srcAccount int32, requiredConfirmations int32, sendAll bool) (*UnsignedTransaction, error) {
+	// output destination
+	addr, err := dcrutil.DecodeAddress(destAddr)
+	if err != nil {
+		return nil, err
+	}
+	pkScript, err := txscript.PayToAddrScript(addr)
+	if err != nil {
+		log.Error(err)
+		return nil, err
+	}
+	version := txscript.DefaultScriptVersion
+
+	// pay output
+	outputs := make([]*wire.TxOut, 0)
+	var algo wallet.OutputSelectionAlgorithm = wallet.OutputSelectionAlgorithmAll
+	var changeSource txauthor.ChangeSource
+	if !sendAll {
+		algo = wallet.OutputSelectionAlgorithmDefault
+		output := &wire.TxOut{
+			Value:    amount,
+			Version:  version,
+			PkScript: pkScript,
+		}
+		outputs = append(outputs, output)
+	} else {
+		changeSource, err = txhelper.MakeTxChangeSource(destAddr)
+		if err != nil {
+			log.Error(err)
+			return nil, err
+		}
+	}
+	feePerKb := txrules.DefaultRelayFeePerKb
+
+	// create tx
+	tx, err := lw.wallet.NewUnsignedTransaction(outputs, feePerKb, uint32(srcAccount),
+		requiredConfirmations, algo, changeSource)
+	if err != nil {
+		log.Error(err)
+		return nil, translateError(err)
+	}
+
+	if tx.ChangeIndex >= 0 {
+		tx.RandomizeChangePosition()
+	}
+
+	var txBuf bytes.Buffer
+	txBuf.Grow(tx.Tx.SerializeSize())
+	err = tx.Tx.Serialize(&txBuf)
+	if err != nil {
+		log.Error(err)
+		return nil, err
+	}
+
+	var totalOutput dcrutil.Amount
+	for _, txOut := range outputs {
+		totalOutput += dcrutil.Amount(txOut.Value)
+	}
+
+	return &UnsignedTransaction{
+		UnsignedTransaction:       txBuf.Bytes(),
+		TotalOutputAmount:         int64(totalOutput),
+		TotalPreviousOutputAmount: int64(tx.TotalInput),
+		EstimatedSignedSize:       tx.EstimatedSignedSerializeSize,
+		ChangeIndex:               tx.ChangeIndex,
+	}, nil
+}
+
+func (lw *LibWallet) SendTransaction(privPass []byte, destAddr string, amount int64, srcAccount int32, requiredConfs int32, sendAll bool) ([]byte, error) {
+	n, err := lw.wallet.NetworkBackend()
+	if err != nil {
+		log.Error(err)
+		return nil, err
+	}
+	defer func() {
+		for i := range privPass {
+			privPass[i] = 0
+		}
+	}()
+	// output destination
+	addr, err := dcrutil.DecodeAddress(destAddr)
+	if err != nil {
+		return nil, err
+	}
+	pkScript, err := txscript.PayToAddrScript(addr)
+	if err != nil {
+		log.Error(err)
+		return nil, err
+	}
+
+	// pay output
+	outputs := make([]*wire.TxOut, 0)
+	var algo wallet.OutputSelectionAlgorithm = wallet.OutputSelectionAlgorithmAll
+	var changeSource txauthor.ChangeSource
+	if !sendAll {
+		algo = wallet.OutputSelectionAlgorithmDefault
+		output := &wire.TxOut{
+			Value:    amount,
+			Version:  txscript.DefaultScriptVersion,
+			PkScript: pkScript,
+		}
+		outputs = append(outputs, output)
+	} else {
+		changeSource, err = txhelper.MakeTxChangeSource(destAddr)
+		if err != nil {
+			log.Error(err)
+			return nil, err
+		}
+	}
+
+	// create tx
+	unsignedTx, err := lw.wallet.NewUnsignedTransaction(outputs, txrules.DefaultRelayFeePerKb, uint32(srcAccount),
+		requiredConfs, algo, changeSource)
+	if err != nil {
+		log.Error(err)
+		return nil, translateError(err)
+	}
+
+	if unsignedTx.ChangeIndex >= 0 {
+		unsignedTx.RandomizeChangePosition()
+	}
+
+	var txBuf bytes.Buffer
+	txBuf.Grow(unsignedTx.Tx.SerializeSize())
+	err = unsignedTx.Tx.Serialize(&txBuf)
+	if err != nil {
+		log.Error(err)
+		return nil, err
+	}
+
+	var tx wire.MsgTx
+	err = tx.Deserialize(bytes.NewReader(txBuf.Bytes()))
+	if err != nil {
+		log.Error(err)
+		//Bytes do not represent a valid raw transaction
+		return nil, err
+	}
+
+	lock := make(chan time.Time, 1)
+	defer func() {
+		lock <- time.Time{}
+	}()
+
+	err = lw.wallet.Unlock(privPass, lock)
+	if err != nil {
+		log.Error(err)
+		return nil, errors.New(ErrInvalidPassphrase)
+	}
+
+	var additionalPkScripts map[wire.OutPoint][]byte
+
+	invalidSigs, err := lw.wallet.SignTransaction(&tx, txscript.SigHashAll, additionalPkScripts, nil, nil)
+	if err != nil {
+		log.Error(err)
+		return nil, err
+	}
+
+	invalidInputIndexes := make([]uint32, len(invalidSigs))
+	for i, e := range invalidSigs {
+		invalidInputIndexes[i] = e.InputIndex
+	}
+
+	var serializedTransaction bytes.Buffer
+	serializedTransaction.Grow(tx.SerializeSize())
+	err = tx.Serialize(&serializedTransaction)
+	if err != nil {
+		log.Error(err)
+		return nil, err
+	}
+
+	var msgTx wire.MsgTx
+	err = msgTx.Deserialize(bytes.NewReader(serializedTransaction.Bytes()))
+	if err != nil {
+		//Invalid tx
+		log.Error(err)
+		return nil, err
+	}
+
+	txHash, err := lw.wallet.PublishTransaction(&msgTx, serializedTransaction.Bytes(), n)
+	if err != nil {
+		return nil, translateError(err)
+	}
+	return txHash[:], nil
+}
+
+func (lw *LibWallet) PublishUnminedTransactions() error {
+	netBackend, err := lw.wallet.NetworkBackend()
+	if err != nil {
+		return errors.New(ErrNotConnected)
+	}
+	ctx, _ := contextWithShutdownCancel(context.Background())
+	err = lw.wallet.PublishUnminedTransactions(ctx, netBackend)
+	return err
+}
