+++ utils.go
@@ -1,0 +1,76 @@
+package dcrlibwallet
+
+import (
+	"context"
+	"encoding/base64"
+	"encoding/hex"
+	"os"
+	"os/signal"
+
+	"github.com/decred/dcrd/dcrutil"
+)
+
+func shutdownListener() {
+	interruptChannel := make(chan os.Signal, 1)
+	signal.Notify(interruptChannel, signals...)
+
+	// Listen for the initial shutdown signal
+	select {
+	case sig := <-interruptChannel:
+		log.Infof("Received signal (%s).  Shutting down...", sig)
+	case <-shutdownRequestChannel:
+		log.Info("Shutdown requested.  Shutting down...")
+	}
+
+	// Cancel all contexts created from withShutdownCancel.
+	close(shutdownSignaled)
+
+	// Listen for any more shutdown signals and log that shutdown has already
+	// been signaled.
+	for {
+		select {
+		case <-interruptChannel:
+		case <-shutdownRequestChannel:
+		}
+		log.Info("Shutdown signaled.  Already shutting down...")
+	}
+}
+
+func contextWithShutdownCancel(ctx context.Context) (context.Context, context.CancelFunc) {
+	ctx, cancel := context.WithCancel(ctx)
+	go func() {
+		<-shutdownSignaled
+		cancel()
+	}()
+	return ctx, cancel
+}
+
+func AmountCoin(amount int64) float64 {
+	return dcrutil.Amount(amount).ToCoin()
+}
+
+func AmountAtom(f float64) int64 {
+	amount, err := dcrutil.NewAmount(f)
+	if err != nil {
+		log.Error(err)
+		return -1
+	}
+	return int64(amount)
+}
+
+func EncodeHex(hexBytes []byte) string {
+	return hex.EncodeToString(hexBytes)
+}
+
+func EncodeBase64(text []byte) string {
+	return base64.StdEncoding.EncodeToString(text)
+}
+
+func DecodeBase64(base64Text string) ([]byte, error) {
+	b, err := base64.StdEncoding.DecodeString(base64Text)
+	if err != nil {
+		return nil, err
+	}
+
+	return b, nil
+}
