+++ utils/utils.go
@@ -1,0 +1,65 @@
+package utils
+
+import (
+	"encoding/base64"
+	"net"
+
+	"github.com/decred/dcrd/hdkeychain"
+	"github.com/decred/dcrwallet/walletseed"
+	"time"
+)
+
+func NormalizeAddress(addr string, defaultPort string) (string, error) {
+	// If the first SplitHostPort errors because of a missing port and not
+	// for an invalid host, add the port.  If the second SplitHostPort
+	// fails, then a port is not missing and the original error should be
+	// returned.
+	host, port, origErr := net.SplitHostPort(addr)
+	if origErr == nil {
+		return net.JoinHostPort(host, port), nil
+	}
+	addr = net.JoinHostPort(addr, defaultPort)
+	_, _, err := net.SplitHostPort(addr)
+	if err != nil {
+		return "", origErr
+	}
+	return addr, nil
+}
+
+func GenerateSeed() ([]byte, string, error) {
+	seed, err := hdkeychain.GenerateSeed(hdkeychain.RecommendedSeedLen)
+	if err != nil {
+		return nil, "", err
+	}
+
+	return seed, walletseed.EncodeMnemonic(seed), nil
+}
+
+func VerifySeed(seedMnemonic string) bool {
+	_, err := walletseed.DecodeUserInput(seedMnemonic)
+	return err == nil
+}
+
+func DecodeBase64(base64Text string) ([]byte, error) {
+	b, err := base64.StdEncoding.DecodeString(base64Text)
+	if err != nil {
+		return nil, err
+	}
+
+	return b, nil
+}
+
+// ExtractDateOrTime returns the date represented by the timestamp as a date string if the timestamp is over 24 hours ago.
+// Otherwise, the time alone is returned as a string.
+func ExtractDateOrTime(timestamp int64) string {
+	utcTime := time.Unix(timestamp, 0).UTC()
+	if time.Now().UTC().Sub(utcTime).Hours() > 24 {
+		return utcTime.Format("2006-01-02")
+	} else {
+		return utcTime.Format("15:04:05")
+	}
+}
+
+func FormatUTCTime(timestamp int64) string {
+	return time.Unix(timestamp, 0).UTC().Format("2006-01-02 15:04:05")
+}
